dodana mogucnost dodavanja koordinata lokacije gubitka ljubimca

+++ IzvorniKod/backend/src/main/java/com/lostpetfinder/entity/County.java
@@ -3,7 +3,7 @@ package com.lostpetfinder.entity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Size;
 
-/*
+
 @Entity
 @Table(name = "counties")
 public class County {
@@ -11,7 +11,7 @@ public class County {
     // not sure if counties even have their own assigned value, so I'm just generating it like this for now
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int countyId;
+    private Long countyId;
 
     // potentially update the min attribute from @Size
     @Column(nullable = false)
@@ -24,7 +24,7 @@ public class County {
         this.name = name;
     }
 
-    public int getCountyId() {
+    public Long getCountyId() {
         return countyId;
     }
 
@@ -36,4 +36,3 @@ public class County {
         this.name = name;
     }
 }
-*/
